Add exercise event form into modal and add css
