+++ frontend/src/components/Exercise/ExerciseEventForm.css
@@ -1,0 +1,112 @@
+.exercise-form-container {
+    display: flex;
+    flex-direction: column;
+	justify-content: center;
+	align-items: center;
+	background-color: var(--form-background-color);
+    padding: 0px 50px;
+	/* height: 80vh; */
+}
+
+.exercise-form {
+	border-radius: 8px;
+	padding: 30px;
+	animation: exercise-form-border 0.8s ease-in 0s 1 normal forwards;
+	display: flex;
+	flex-direction: column;
+	justify-content: space-around;
+	font-size: 1em;
+}
+
+.exercise-form > * {
+	margin: 10px 0px;
+	font-size: 1em;
+}
+
+.exercise-form-container > h2 {
+	align-self: center;
+	font-size: 3rem;
+	font-weight: 700;
+	letter-spacing: 2.5px;
+	color: var(--form-header-color);
+}
+
+#exercise-input-span, .exercise-entry-form > span {
+    margin-right: 5px;
+    margin-left: 15px;
+    font-weight: bold;
+    color: var(--form-text-input-color);
+}
+
+#exercise-input-div {
+    margin: 15px 0px;
+}
+
+.exercise-input-delete-btn {
+	border: 0px solid var(--form-button-color);
+	background-color: var(--form-button-color);
+	border-radius: 20px;
+	width: 50px;
+	padding: 5px;
+	color: var(--form-text-input-color);
+    margin-left: 10px;
+}
+
+.exercise-entry-form > input[type=textarea] {
+    min-height: 50px;
+    min-width: 475px;
+    margin-top: 5px;
+    border-radius: 8px;
+}
+
+.exercise-entry-form > input[type=number] {
+    height: 30px;
+    width: 30px;
+    margin-top: 5px;
+    border-radius: 8px;
+}
+
+.exercise-entry-form > input[type=date] {
+    height: 30px;
+    border-radius: 8px;
+}
+
+#exercise-input-div > input[type=text] {
+    height: 30px;
+    min-width: 100px;
+    border-radius: 8px;
+}
+
+#exercise-input-div > input[type=number] {
+    height: 30px;
+    border-radius: 8px;
+    width: 100px;
+}
+
+.exercise-event-form-btn-container {
+    display: flex;
+    justify-content: space-around;
+}
+
+.exercise-event-form-btn {
+    width: 100%;
+    border: 0px solid var(--form-button-color);
+	background-color: var(--form-button-color);
+	color: var(--form-text-input-color);
+	border-radius: 20px;
+	padding: 5px;
+    margin: 0 5px;
+}
+
+.exercise-input-delete-btn:hover {
+    cursor: pointer;
+}
+
+.exercise-event-form-btn:hover {
+    cursor: pointer;
+}
+
+/* idk how i feel about the HR -.- */
+hr {
+    margin-bottom: 30px;
+}
