+++ frontend/src/components/FloatingMenu/FloatingMenu.js
@@ -2,7 +2,7 @@ import './FloatingMenu.css';
 import { useSelector } from 'react-redux';
 import { useState } from 'react';
 import { Modal } from '../../context/Modal';
-import ExerciseCreateForm from '../Exercise/ExerciseCreateForm';
+import ExerciseEventForm from '../Exercise/ExerciseEventForm';
 
 // Just to test modal popup
 import LoginForm from '../SessionForms/LoginForm';
@@ -21,8 +21,8 @@ const FloatingMenu = (props) => {
 				</div>
 			}
 			{showExerciseEntry && <Modal onClose={e => setShowExerciseEntry(false)}>
-				<LoginForm />
-				{/* <ExerciseCreateForm /> */}
+				{/* <LoginForm /> */}
+				<ExerciseEventForm />
 			</Modal>}
 		</>
 	)
